Added saving failed message

+++ libs/tinyexr/tinyexr.h
@@ -2436,7 +2436,7 @@ static bool hufBuildDecTable(const long long *hcode,  // i : encoding table
         unsigned int *p = pl->p;
         pl->p = new unsigned int[pl->lit];
 
-        for (int i = 0; i < pl->lit - 1; ++i) pl->p[i] = p[i];
+        for (int i = 0; i < (int)pl->lit - 1; ++i) pl->p[i] = p[i];
 
         delete[] p;
       } else {
@@ -2693,7 +2693,7 @@ static bool hufDecode(const long long *hcode,  // i : encoding table
 
         int j;
 
-        for (j = 0; j < pl.lit; j++) {
+        for (j = 0; j < (int)pl.lit; j++) {
           int l = hufLength(hcode[pl.p[j]]);
 
           while (lc < l && in < ie)  // get more bits
